moved CouchDb code to Persistence.CouchDb project
