+++ Fabric.Authorization.API/Bootstrapper.cs
@@ -7,10 +7,11 @@ using Fabric.Authorization.API.Extensions;
 using Fabric.Authorization.API.Infrastructure;
 using Fabric.Authorization.API.Infrastructure.PipelineHooks;
 using Fabric.Authorization.API.Logging;
-using Fabric.Authorization.API.Services;
 using Fabric.Authorization.Domain.Events;
 using Fabric.Authorization.Domain.Stores;
 using Fabric.Authorization.Domain.Stores.CouchDB;
+using Fabric.Authorization.Persistence.CouchDb.Configuration;
+using Fabric.Authorization.Persistence.CouchDb.Services;
 using Fabric.Authorization.Persistence.InMemory.Services;
 using Fabric.Authorization.Persistence.SqlServer.Services;
 using Fabric.Platform.Bootstrappers.Nancy;
