+++ Fabric.Authorization.API/Configuration/IAppConfiguration.cs
@@ -1,4 +1,5 @@
 ﻿using Fabric.Authorization.Persistence.SqlServer.Configuration;
+﻿using Fabric.Authorization.Persistence.CouchDb.Configuration;
 using Fabric.Platform.Shared.Configuration;
 
 namespace Fabric.Authorization.API.Configuration
