+++ Fabric.Authorization.API/Extensions/TinyIocExtensions.cs
@@ -6,8 +6,9 @@ using Fabric.Authorization.API.Services;
 using Fabric.Authorization.Domain.Resolvers.Permissions;
 using Fabric.Authorization.Domain.Services;
 using Fabric.Authorization.Domain.Stores;
-using Fabric.Authorization.Domain.Stores.CouchDB;
 using Fabric.Authorization.Domain.Stores.Services;
+using Fabric.Authorization.Persistence.CouchDb.Services;
+using Fabric.Authorization.Persistence.CouchDb.Stores;
 using Fabric.Authorization.Persistence.InMemory.Stores;
 using Fabric.Authorization.Persistence.SqlServer.Services;
 using Fabric.Authorization.Persistence.SqlServer.Stores;
