+++ Fabric.Authorization.API/Startup.cs
@@ -6,7 +6,8 @@ using Fabric.Authorization.API.Constants;
 using Fabric.Authorization.API.Infrastructure;
 using Fabric.Authorization.API.Services;
 using Fabric.Authorization.Domain.Stores;
-using Fabric.Authorization.Domain.Stores.CouchDB;
+using Fabric.Authorization.Persistence.CouchDb.Services;
+using Fabric.Authorization.Persistence.CouchDb.Stores;
 using Fabric.Authorization.Persistence.InMemory.Stores;
 using Fabric.Platform.Auth;
 using Fabric.Platform.Logging;
