+++ Fabric.Authorization.Persistence.CouchDb/Stores/CouchDBPermissionStore.cs
@@ -4,9 +4,11 @@ using System.Threading.Tasks;
 using Fabric.Authorization.Domain.Exceptions;
 using Fabric.Authorization.Domain.Models;
 using Fabric.Authorization.Domain.Services;
+using Fabric.Authorization.Domain.Stores;
+using Fabric.Authorization.Persistence.CouchDb.Services;
 using Serilog;
 
-namespace Fabric.Authorization.Domain.Stores.CouchDB
+namespace Fabric.Authorization.Persistence.CouchDb.Stores
 {
     public class CouchDbPermissionStore : FormattableIdentifierStore<Guid, Permission>, IPermissionStore
     {
