+++ Fabric.Authorization.Persistence.CouchDb/Stores/CouchDbUserStore.cs
@@ -1,9 +1,11 @@
 ﻿using System.Threading.Tasks;
 using Fabric.Authorization.Domain.Models;
 using Fabric.Authorization.Domain.Services;
+using Fabric.Authorization.Domain.Stores;
+using Fabric.Authorization.Persistence.CouchDb.Services;
 using Serilog;
 
-namespace Fabric.Authorization.Domain.Stores.CouchDB
+namespace Fabric.Authorization.Persistence.CouchDb.Stores
 {
     public class CouchDbUserStore : FormattableIdentifierStore<string, User>, IUserStore
     {
