+++ Fabric.Authorization.Persistence.CouchDb/Stores/FormattableIdentifierStore.cs
@@ -1,8 +1,11 @@
 ﻿using Fabric.Authorization.Domain.Models;
 using Fabric.Authorization.Domain.Services;
+using Fabric.Authorization.Domain.Stores;
+using Fabric.Authorization.Domain.Stores.CouchDB;
+using Fabric.Authorization.Persistence.CouchDb.Services;
 using Serilog;
 
-namespace Fabric.Authorization.Domain.Stores.CouchDB
+namespace Fabric.Authorization.Persistence.CouchDb.Stores
 {
     public abstract class FormattableIdentifierStore<K, T> : CouchDbGenericStore<K, T>, IFormattableIdentifier
         where T : ITrackable, IIdentifiable
